Implementação inscrever e desinscrever atividade

+++ migrations/20180918015321-create-customer-role.js
@@ -6,15 +6,15 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.BIGINT
+        type: Sequelize.INTEGER
       },
       customerId: {
-        type: Sequelize.BIGINT,
+        type: Sequelize.INTEGER,
         references: {model: 'customers', key: 'id'},
         allowNull: false
       },
       roleId: {
-        type: Sequelize.BIGINT,
+        type: Sequelize.INTEGER,
         references: {model: 'roles', key: 'id'},
         allowNull: false
       },
